refactor: tree part of journal
